feature: output for session class working
